restructured files whcih had more than 5 fucntions, added some new files
