+++ src/1_input_manager/1_0_tokenizer/tokenizer.c
@@ -6,26 +6,12 @@
 /*   By: anshovah <anshovah@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/21 12:11:14 by anshovah          #+#    #+#             */
-/*   Updated: 2023/11/02 14:44:55 by anshovah         ###   ########.fr       */
+/*   Updated: 2023/11/02 19:12:49 by anshovah         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"     
-
 	
-void print_tokens(t_mbox *mbox)
-{	
-	t_token *current;	
-	
-	current = mbox->tokens;	
-	while (current)	
-	{	
-		printf ("type:(%d) \t token:(%s)\n", current->type, current->value);	
-		current = current->next;	
-	}	
-}	
-	
-
 /**
  * @brief	accepts a string which could contain shifted values and cheks if:
  * 				- contains only one char
@@ -89,31 +75,20 @@ static void	add_token(t_mbox *mbox, char *value, int token_type)
 		cur->next = new_t;
 	}
 }
-/*
-	accepts a not shifted character, shifts it back to a correct ASCII value
-	and hten compares to the separating charactes and returns a respective
-	token type (enum e_token_type value) 
-*/
-int	get_token_type(char c)
-{
-	if (remove_offset(c) == '|')
-		return (PIPE_TOKEN);
-	else if(remove_offset(c) == '<')
-		return (RED_IN_TOKEN);
-	else if(remove_offset(c) == '>')
-		return (RED_OUT_TOKEN);
-	return (WORD_TOKEN);		
-}
 
-/*
-	called in a case if just a sequence of alphanumeric characters without any 
-	separators needs to be transformen into a token. uses call by reference
-	to update int i in a calling function to shift the string correctly
-*/
+/**
+ * @brief	called in a case if just a sequence of alphanumeric characters
+ * 			without any separators needs to be transformen into a token.
+ * 			uses call by reference to update int i in a calling function
+ * 			to shift the string correctly
+ * 
+ * @param	mbox 
+ * @param	str 
+ * @param	i 
+ */
 static void	just_word(t_mbox *mbox, char *str, int *i)
 {
 	int	j;
-	// heloo?????|????wc
 	if (ft_issep(remove_offset(*str)))
 	{
 		add_token(mbox, ft_substr(str, 0, 1), get_token_type(*str));
@@ -124,8 +99,6 @@ static void	just_word(t_mbox *mbox, char *str, int *i)
 		j = 0;
 		while (str[j])
 		{
-			// if (ft_issep(remove_offset(str[j]))
-			// 	|| ft_isqoute(remove_offset(str[j]))) 
 			if (ft_issep(remove_offset(str[j]))) 
 				break ;
 			j++;
@@ -135,10 +108,16 @@ static void	just_word(t_mbox *mbox, char *str, int *i)
 	}
 }
 
-/*
-	receives the substing with a special character between words,
-	splits them into words and characters and adds new tokens to mbox->tokens
-*/
+/**
+ * @brief	receives the substing with a special character between words, 
+ * 			splits them into words and characters and adds new tokens 
+ * 			to mbox->tokens
+ * 
+ * @param	mbox 
+ * @param	str 
+ * @param	i 
+ * @param	quote_state 
+ */
 static void	split_by_sep(t_mbox *mbox, char *str, int i, int quote_state)
 {
 	while (*str)
@@ -162,53 +141,10 @@ static void	split_by_sep(t_mbox *mbox, char *str, int i, int quote_state)
 			if (quote_state != OUT_Q)
 				add_token(mbox, ft_substr(str, 0, i), get_token_type(*str));
 			str += i;
-
-			// FIXME: JUST CHANGE THAT QOUTES DONT START A NEW TOKEN!
 		}
 	}
 }
 
-void	print_tokenizer_output(t_mbox *mbox)
-{
-	printf("\n ------------------------------------ \n");
-	printf("|           TOKENIZER                |\n");
-	printf(" ------------------------------------ \n");
-	print_tokens(mbox);
-	printf(" ------------------------------------ \n");
-}
-
-/**
- * @brief	cheeck spreadsheet!
- * 			if the last char 	of str1 is a >
- * 			and the first chr 	of str2 is a >
- * 			then we have something like "echo hi > > lol"
- * 			-> this is an error so exit cycle
- * 
- * @return t_bool 
- */
-static t_bool	check_special_case_redir_symbol(char *str1, char *str2)
-{
-	t_bool	is_correct;
-
-	is_correct = ft_true;
-	if (!str1 || !str2)
-		return (is_correct);
-		
-	if (str1[ft_strlen(str1) - 1] == add_offset('>')
-		&& str2[0] == add_offset('>'))
-	{
-		create_error_msg("n", "syntax error near unexpected token `>'");
-		is_correct = ft_false;
-	}
-
-	if (str1[ft_strlen(str1) - 1] == add_offset('<')
-		&& str2[0] == add_offset('<'))
-	{
-		create_error_msg("n", "syntax error near unexpected token `<'");
-		is_correct = ft_false;
-	}
-	return (is_correct);
-}
 
 /**
  * @brief	creates a linked list (ll) with tokens generated from
@@ -234,19 +170,13 @@ t_bool	tokenize(t_mbox *mbox, int i)
 	no_space = ft_split(mbox->inp_expand, NO_SPACE);
 	while (no_space[i])
 	{
-		if (!check_special_case_redir_symbol(no_space[i], no_space[i+1]))
+		if (!check_space_between_redir(no_space[i], no_space[i+1]))
 		{
 			free_whatever("m", no_space);
 			free_cycle_v2(mbox);
 		}
-		if (ft_strchr(no_space[i], add_offset('|'))
-			|| ft_strchr(no_space[i], add_offset('>'))
-			|| ft_strchr(no_space[i], add_offset('<'))
-			|| ft_strchr(no_space[i], add_offset('\''))
-			|| ft_strchr(no_space[i], add_offset('"')))
-		{
+		if (check_sp(no_space[i]))
 			split_by_sep(mbox, no_space[i], 0, OUT_Q);
-		}
 		else
 			add_token(mbox, ft_strdup(no_space[i]), WORD_TOKEN);
 		i++;
